refactor(core): split the ERModel into state/run/persist layers
`core/resolver.py` was the largest file in the repo (2,330 lines) and was
essentially one class: ERModel started at :355 and ran ~1,975 lines. It is now
the leaf of a linear chain, each layer owning one responsibility:

  _model_state.py    what a model IS      slots, identity, the 3 construction
                                          doors, schema binding
  _model_run.py      how it RUNS          block -> (compare) -> score -> cluster,
                                          plus dedupe/compare
  _model_persist.py  how it PERSISTS      resolver.json manifest + sidecars
  _artifacts.py      the adapter          component <-> ComponentSpec (a leaf,
                                          no ERModel knowledge)
  resolver.py        ERModel              composition + the backend-binding
                                          surface

A base chain, not free-standing mixins: a mixin annotating `blocker:
Blocker[Any]` would collide with the property implementing it, so each layer
subclasses the one below and inherits the slots as non-Optional types.

Behaviour is unchanged by construction: every moved method was extracted from
d63443a's exact bytes at machine-verified boundaries, not retyped.

Two seams were refused on measured evidence, not taste:

* The fit family (645 lines) stays. `_fit_finetune` lazily imports
  `core.finetune`/`core.matchers.llm_judge`, and every member of resolver's
  9-module SCC reaches back via `openrouter -> benchmark -> resolver`. Hosting
  fit in a new module measures tangled 23 -> 24, largest SCC 9 -> 10: the same
  "a split spread the knot across more modules" regression PR #169 shipped
  unnoticed and tests/test_import_tangle.py exists to catch. Same for the
  from_schema builders and the anchor surface, which is why those stay too.
* Each candidate seam was probed against the real graph BEFORE any code moved.
  persist/artifacts, run/front-door and fit-without-finetune measured neutral;
  the other three measured +1. Only the neutral ones were taken.

`load` splits at the one honest seam: _read_artifact (mechanism: validate,
map specs to slots, rebuild from the registry) lives in the persistence layer,
while choosing the class to instantiate stays at the leaf -- that is the only
part needing the ERModel name, and importing it back would re-knot the graph.

anchor_store reached into resolver for the private _iter_vector_blockers; it now
imports the same walk from _model_run (the anchor_store <-> resolver pair it
closes is unaffected).

Measured, before -> after:
  resolver.py               2,330 -> 1,256 lines
  langres.core.__all__         76 -> 76
  runtime SCC                  [] -> []          (a bare import langres: no cycle)
  all-edges SCCs      [9,9,3,2]/23 -> [9,9,3,2]/23
  core coverage             97.91% -> 97.92%     (same not-slow filter)
  tests                       3643 -> 3643 passing

No gate file was edited.
