+++ tests/test_resolver_config_dict.py
@@ -155,7 +155,7 @@ def test_config_dict_recipe_id_stable_across_version_bump(
     artifact-schema bump would silently fork ``recipe_id`` and break idempotent
     replay. Version/provenance live on the RunContext, unhashed.
     """
-    from langres.core import resolver as resolver_mod
+    from langres.core import _model_persist as persist_mod
     from langres.core.runs import RunContext, compute_recipe_id
 
     resolver = _company_resolver()
@@ -171,13 +171,16 @@ def test_config_dict_recipe_id_stable_across_version_bump(
     baseline = _recipe_id()
 
     # Simulate a package release + artifact-schema bump. Both constants are
-    # patched on `resolver_mod` -- where they are *looked up* -- because that is
-    # where `_build_manifest` reads them from. Patching `langres.__version__`
-    # instead would no-op silently (resolver binds the version from the
+    # patched on `langres.core._model_persist` -- where they are *looked up* --
+    # because that is where `_build_manifest` reads them from. (It read them from
+    # `core.resolver` until the model layer was split by responsibility; the
+    # persistence half, and with it the version envelope, moved there. The patch
+    # target follows the lookup, always.) Patching `langres.__version__` instead
+    # would no-op silently (the manifest binds the version from the
     # `langres._version` leaf, not through the root) and this test would pass
     # while simulating nothing.
-    monkeypatch.setattr(resolver_mod, "LANGRES_VERSION", "999.999.999")
-    monkeypatch.setattr(resolver_mod, "ARTIFACT_VERSION", "999")
+    monkeypatch.setattr(persist_mod, "LANGRES_VERSION", "999.999.999")
+    monkeypatch.setattr(persist_mod, "ARTIFACT_VERSION", "999")
 
     assert _recipe_id() == baseline
